requirements.txt
altair==6.0.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibm-cos-sdk==2.14.3
ibm-cos-sdk-core==2.14.3
ibm-cos-sdk-s3transfer==2.14.3
ibm_watsonx_ai==1.4.11
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lomond==0.3.3
MarkupSafe==3.0.3
narwhals==2.15.0
numpy==2.4.0
packaging==25.0
pandas==2.2.3
pillow==12.1.0
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
streamlit==1.52.2
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uv==0.9.22
watchdog==6.0.0
